move create_coordination_table
