+++ massgen/coordination_tracker.py
@@ -603,7 +603,7 @@ class CoordinationTracker:
         """Generate coordination table using the create_coordination_table.py module."""
         try:
             # Import the table builder
-            from create_coordination_table import CoordinationTableBuilder
+            from massgen.frontend.create_coordination_table import CoordinationTableBuilder
             
             # Create the event-driven table directly from session data (includes metadata)
             builder = CoordinationTableBuilder(session_data)
